Move helper methods from forecasts.js to a helpers folder
