+++ helpers/format_helper.js
@@ -1,0 +1,13 @@
+function formatFullForecast(location, forecast) {
+  var fullForecast = {
+    location: location,
+    currently: forecast.currently,
+    hourly: forecast.hourly,
+    daily: forecast.daily
+  };
+  return fullForecast;
+}
+
+module.exports = {
+  formatFullForecast
+}
